+++ dragonfly-client/src/grpc/interceptor.rs
@@ -1,7 +1,21 @@
-use {
-    tonic::{metadata, service::Interceptor, Request, Status},
-    tracing_opentelemetry::OpenTelemetrySpanExt,
-};
+/*
+ *     Copyright 2024 The Dragonfly Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use tonic::{metadata, service::Interceptor, Request, Status};
+use tracing_opentelemetry::OpenTelemetrySpanExt;
 
 /// MetadataMap is a tracing meda data map container.
 struct MetadataMap<'a>(&'a mut metadata::MetadataMap);
