+++ dragonfly-client/src/grpc/mod.rs
@@ -27,9 +27,9 @@ use tracing::{error, info, instrument, Instrument};
 pub mod dfdaemon_download;
 pub mod dfdaemon_upload;
 pub mod health;
+pub mod interceptor;
 pub mod manager;
 pub mod scheduler;
-pub mod tracing_grpc;
 
 /// CONNECT_TIMEOUT is the timeout for GRPC connection.
 pub const CONNECT_TIMEOUT: Duration = Duration::from_secs(2);
